Refactor calculation CantPureza location
