+++ CrianzaMonturas/CrianzaMonturas.Core/Vista/CrearCria.cs
@@ -1,16 +1,21 @@
-﻿using CrianzaMonturas.Dal.Modelo;
+﻿using CrianzaMonturas.Dal.Contratos;
+using CrianzaMonturas.Dal.Modelo;
 
 namespace CrianzaMonturas.Core.Vista
 {
     public partial class CrearCria : Form
     {
+        #region Properties
+
         private string nombre;
         private string sexo;
         private bool predispuesto;
+        private int pureza;
 
-        public string Nombre {
+        public string Nombre
+        {
             get { return nombre; }
-            set { nombre = value; } 
+            set { nombre = value; }
         }
 
         public string Sexo
@@ -25,16 +30,27 @@ namespace CrianzaMonturas.Core.Vista
             set { predispuesto = value; }
         }
 
-        public CrearCria(Tipo tipo)
+        public int Pureza
+        {
+            get { return pureza; }
+            set { pureza = value; }
+        }
+
+        #endregion Properties
+
+        public CrearCria(Tipo tipo, Montura padre, Montura madre)
         {
             InitializeComponent();
-            
+
             nombre = tipo.Nombre;
             sexo = string.Empty;
 
             txtTipo.Text = nombre;
             txtNombre.Text = tipo.Sigla;
+            txtPureza.Text = CalcularPureza(tipo.Id, padre, madre, 1).ToString();
         }
+
+        #region Events
 
         private void BtnCrear_Click(object sender, EventArgs e)
         {
@@ -47,6 +63,7 @@ namespace CrianzaMonturas.Core.Vista
                 Nombre = txtNombre.Text;
                 Sexo = cmbSexo.Text;
                 Predispuesto = chkPredispuesto.Checked;
+                Pureza = Convert.ToInt32(txtPureza.Text);
                 this.DialogResult = DialogResult.OK;
             }
             else
@@ -61,5 +78,51 @@ namespace CrianzaMonturas.Core.Vista
         {
             Clipboard.SetText(txtNombre.Text);
         }
+
+
+        private void ChkModificar_CheckedChanged(object sender, EventArgs e)
+        {
+            txtPureza.ReadOnly = chkModificar.Checked;
+        }
+
+        private void TxtPureza_KeyPress(object sender, KeyPressEventArgs e)
+        {
+            if (!char.IsControl(e.KeyChar) && !char.IsDigit(e.KeyChar))
+            {
+                e.Handled = true;
+            }
+        }
+
+        #endregion Events
+
+        #region Methods
+
+        private int CalcularPureza(int tipoCria, IMontura? padre, IMontura? madre, int lvl)
+        {
+            var pureza = 0;
+            var base_pureza = lvl switch
+            {
+                1 => 6,
+                2 => 3,
+                3 => 1,
+                _ => 0,
+            };
+
+            if (padre is not null && padre.TipoId == tipoCria) pureza += base_pureza;
+            if (madre is not null && madre.TipoId == tipoCria) pureza += base_pureza;
+
+            if (lvl <= 2)
+            {
+                var newLvl = lvl + 1;
+                if (padre is not null) pureza += CalcularPureza(tipoCria, padre.Padre, padre.Madre, newLvl);
+                if (madre is not null) pureza += CalcularPureza(tipoCria, madre.Padre, madre.Madre, newLvl);
+            }
+
+            return pureza;
+        }
+
+        #endregion Methods
+
+        
     }
 }
